Move avl and spl_list walkers from zfs subdir to spl
= Motivation

Now that we have an SPL folder we want to move there
all the spl commands from the zfs directory.

= Testing

Ensured that the `spa` and `zfs_dbgmsg` commands work
as before (they are consumers of these two walkers).
