+++ sdb/commands/spl/spl_list.py
